Added documentation for ManagedObjectController ReadBodyFilter
